[fix] separated scheduler factory into optimisation folder to keep training loop clean
