+++ leaderboard/leaderboard.csv
@@ -154,4 +154,5 @@ timestamp,run_name,runner,git_commit,model_path,best_epoch,best_metric_name,best
 2026-05-09 20:28:56,safetybert_final_best,Aashutosh Chapagain,c341f6c,trained_models/20260509_202430_safetybert/safetybert_model.pt,6,f1_macro,0.6126974821090698,1.5433651279296554,0.6624472737312317,0.6126974821090698,0.6558790802955627,0.6246506571769714,0.609251856803894,6,2,2e-05,1.0,cross_entropy,False,SafetyBERT,BertClassifier,AdamW,ReduceLROnPlateau,CrossEntropyLoss,True,261.6881980895996,1.4920904560312005,0.6974790096282959,0.6202911734580994,0.6890931129455566,0.6524436473846436,0.6300891637802124,0.8529411764705882,,True,0.8529411764705882,0.1050420168067226,0.042016806722689,,,False,"{""Gravitational"": {""f1"": 0.8351647853851318, ""target"": 0.7, ""met"": true}, ""Human"": {""f1"": 0.6875000596046448, ""target"": 0.7, ""met"": false}, ""Other"": {""f1"": 0.6274510025978088, ""target"": 0.7, ""met"": false}, ""Vehicular"": {""f1"": 0.7999999523162842, ""target"": 0.7, ""met"": true}}",,
 2026-05-09 20:35:19,safetybert_final_best,Aashutosh Chapagain,c341f6c,trained_models/20260509_203011_safetybert/safetybert_model.pt,5,f1_macro,0.6506102085113525,1.3863389039341407,0.6751055121421814,0.6506102085113525,0.6744024753570557,0.6409677863121033,0.6731047034263611,10,2,2e-05,1.0,cross_entropy,False,SafetyBERT,BertClassifier,AdamW,ReduceLROnPlateau,CrossEntropyLoss,True,304.3410828113556,1.308960524927668,0.7100840210914612,0.6459318399429321,0.708693265914917,0.6733250021934509,0.668599009513855,0.7394957983193278,,True,0.7394957983193278,0.2016806722689075,0.0588235294117647,,,False,"{""Gravitational"": {""f1"": 0.7956989407539368, ""target"": 0.7, ""met"": true}, ""Human"": {""f1"": 0.7088607549667358, ""target"": 0.7, ""met"": true}, ""Other"": {""f1"": 0.6440678238868713, ""target"": 0.7, ""met"": false}, ""Vehicular"": {""f1"": 0.8314606547355652, ""target"": 0.7, ""met"": true}}",,
 2026-05-09 20:54:37,safetybert_model2,Aashutosh Chapagain,2817821,trained_models/20260509_204844_safetybert/safetybert_model.pt,6,f1_macro,0.6957728862762451,1.1906913505052217,0.8041666746139526,0.6957728862762451,0.8025131225585938,0.6948038935661316,0.7033060193061829,10,2,2e-05,1.0,cross_entropy,False,SafetyBERT,BertClassifier,AdamW,ReduceLROnPlateau,CrossEntropyLoss,False,348.8565831184387,1.286983242771627,0.7791666388511658,0.6874234676361084,0.7877380847930908,0.6582596302032471,0.7305638194084167,0.9416666666666668,,True,0.9416666666666668,0.0583333333333333,0.0,,,False,"{""class_5"": {""f1"": 0.0, ""target"": 0.7, ""met"": false}, ""class_6"": {""f1"": 0.0, ""target"": 0.7, ""met"": false}, ""class_11"": {""f1"": 0.0, ""target"": 0.7, ""met"": false}, ""class_17"": {""f1"": 0.0, ""target"": 0.7, ""met"": false}}",,
-2026-05-10 17:45:33,safetybert_final_best-repeated,Aashutosh Chapagain,37cf0cf,trained_models/20260510_173748_safetybert/safetybert_model.pt,9,f1_macro,0.6379886269569397,1.8947518923111606,0.6793248653411865,0.6379886269569397,0.677941083908081,0.6476214528083801,0.6411352157592773,10,2,2e-05,1.0,cross_entropy,False,SafetyBERT,BertClassifier,AdamW,ReduceLROnPlateau,CrossEntropyLoss,True,458.91593503952026,1.858884473996503,0.6890756487846375,0.6005118489265442,0.6831992268562317,0.5883564352989197,0.6206754446029663,0.8865546218487395,,True,0.8865546218487395,0.07563025210084033,0.037815126050420166,,,False,"{""Gravitational"": {""f1"": 0.8571428060531616, ""target"": 0.7, ""met"": true}, ""Human"": {""f1"": 0.739130437374115, ""target"": 0.7, ""met"": true}, ""Other"": {""f1"": 0.5714285373687744, ""target"": 0.7, ""met"": false}, ""Vehicular"": {""f1"": 0.7999999523162842, ""target"": 0.7, ""met"": true}}","{""model_type"": ""SafetyBERT"", ""target_type"": ""energy"", ""model_name"": ""adanish91/safetybert"", ""tokenizer_name"": ""bert-base-uncased"", ""pooling"": ""cls"", ""fine_tune"": true, ""batch_size"": 8, ""label_col"": ""energy_type"", ""text_col"": ""description"", ""learning_rate"": 2e-05, ""classifier_dropout"": 0.275, ""embedding_dropout"": 0.1, ""max_length"": 160, ""use_class_weights"": false, ""weight_decay"": 0.01, ""threshold"": 0.8, ""patience"": 2, ""scheduler_config"": {""name"": ""ReduceLROnPlateau"", ""monitor"": ""f1_macro"", ""mode"": ""max"", ""factor"": 0.5, ""patience"": 1, ""min_lr"": 1e-07, ""step_per_batch"": false}, ""device"": ""mps""}",False
+2026-05-10 17:45:33,safetybert_final_best-repeated,Aashutosh Chapagain,37cf0cf,trained_models/20260510_173748_safetybert/safetybert_model.pt,9,f1_macro,0.6379886269569397,1.8947518923111608,0.6793248653411865,0.6379886269569397,0.677941083908081,0.6476214528083801,0.6411352157592773,10,2,2e-05,1.0,cross_entropy,False,SafetyBERT,BertClassifier,AdamW,ReduceLROnPlateau,CrossEntropyLoss,True,458.91593503952026,1.858884473996503,0.6890756487846375,0.6005118489265442,0.6831992268562317,0.5883564352989197,0.6206754446029663,0.8865546218487395,,True,0.8865546218487395,0.0756302521008403,0.0378151260504201,,,False,"{""Gravitational"": {""f1"": 0.8571428060531616, ""target"": 0.7, ""met"": true}, ""Human"": {""f1"": 0.739130437374115, ""target"": 0.7, ""met"": true}, ""Other"": {""f1"": 0.5714285373687744, ""target"": 0.7, ""met"": false}, ""Vehicular"": {""f1"": 0.7999999523162842, ""target"": 0.7, ""met"": true}}","{""model_type"": ""SafetyBERT"", ""target_type"": ""energy"", ""model_name"": ""adanish91/safetybert"", ""tokenizer_name"": ""bert-base-uncased"", ""pooling"": ""cls"", ""fine_tune"": true, ""batch_size"": 8, ""label_col"": ""energy_type"", ""text_col"": ""description"", ""learning_rate"": 2e-05, ""classifier_dropout"": 0.275, ""embedding_dropout"": 0.1, ""max_length"": 160, ""use_class_weights"": false, ""weight_decay"": 0.01, ""threshold"": 0.8, ""patience"": 2, ""scheduler_config"": {""name"": ""ReduceLROnPlateau"", ""monitor"": ""f1_macro"", ""mode"": ""max"", ""factor"": 0.5, ""patience"": 1, ""min_lr"": 1e-07, ""step_per_batch"": false}, ""device"": ""mps""}",False
+2026-05-10 18:17:30,safetybert_final_best-repeated-02,Aashutosh Chapagain,97e5380,trained_models/20260510_181026_safetybert/safetybert_model.pt,7,f1_macro,0.6695519685745239,1.5736697059643419,0.7130801677703857,0.6695519685745239,0.7086651921272278,0.6783162355422974,0.6787891983985901,10,2,2e-05,1.0,cross_entropy,False,SafetyBERT,BertClassifier,AdamW,ReduceLROnPlateau,CrossEntropyLoss,True,418.47955203056335,1.6757062489182508,0.6932772994041443,0.6255591511726379,0.6876384615898132,0.6378886699676514,0.6489014029502869,0.8697478991596639,,True,0.8697478991596639,0.1134453781512605,0.01680672268907563,,,False,"{""Gravitational"": {""f1"": 0.804347813129425, ""target"": 0.7, ""met"": true}, ""Human"": {""f1"": 0.7058823704719543, ""target"": 0.7, ""met"": true}, ""Other"": {""f1"": 0.5714285373687744, ""target"": 0.7, ""met"": false}, ""Vehicular"": {""f1"": 0.804347813129425, ""target"": 0.7, ""met"": true}}","{""model_type"": ""SafetyBERT"", ""target_type"": ""energy"", ""model_name"": ""adanish91/safetybert"", ""tokenizer_name"": ""bert-base-uncased"", ""pooling"": ""cls"", ""fine_tune"": true, ""batch_size"": 8, ""label_col"": ""energy_type"", ""text_col"": ""description"", ""learning_rate"": 2e-05, ""classifier_dropout"": 0.275, ""embedding_dropout"": 0.1, ""max_length"": 160, ""use_class_weights"": false, ""weight_decay"": 0.01, ""threshold"": 0.8, ""patience"": 2, ""scheduler_config"": {""name"": ""ReduceLROnPlateau"", ""monitor"": ""f1_macro"", ""mode"": ""max"", ""factor"": 0.5, ""patience"": 1, ""min_lr"": 1e-07, ""step_per_batch"": false}, ""device"": ""mps""}",False
