+++ modules/optimisation/__init__.py
@@ -1,0 +1,8 @@
+"""Optimisation utilities for optimisers and learning-rate schedulers."""
+
+from .scheduler_factory import normalise_scheduler_config, create_scheduler
+
+__all__ = [
+    "normalise_scheduler_config",
+    "create_scheduler",
+]
