+++ modules/training_loop/config.py
@@ -10,7 +10,7 @@ import torch.optim as optim
 from .utility import _safe_class_name
 from .loss import get_loss_function
 from .imbalance import make_weighted_sampler
-from .scheduler import normalise_scheduler_config, create_scheduler
+from modules.optimisation import normalise_scheduler_config, create_scheduler
 
 
 # CONFIG AND UTILITY FUNCTIONS FOR TRAINING LOOP
